requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
APScheduler==3.0.0
boto3==1.17.79
botocore==1.20.79
dj-database-url==0.5.0
Django==3.2.3
django-easy-pdf3==0.1.2
django-heroku==0.3.1
django-storages==1.11.1
django-tinymce==3.3.0
djangorestframework==3.12.4
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
jmespath==0.10.0
Pillow==8.2.0
psycopg2==2.9.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
reportlab==3.5.68
requests==2.26.0
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
xlsxwriter==1.4.5
